+++ src/RecommendedUtils.hpp
@@ -1,0 +1,13 @@
+#pragma once
+#include <Geode/Geode.hpp>
+
+//geode namespace
+using namespace geode::prelude;
+
+class RecommendedUtils {
+public:
+	static void validateLevels();
+	static void generateRecommendations();
+
+	static bool hasPartial(int);
+};