+++ src/menus/RecommendedLayer.cpp
@@ -1,0 +1,261 @@
+//geode header
+#include <Geode/Geode.hpp>
+
+//other headers
+#include <Geode/utils/web.hpp>
+#include <Geode/loader/Event.hpp>
+#include <Geode/utils/JsonValidation.hpp>
+
+#include "DPLayer.hpp"
+#include "../popups/StatsPopup.hpp"
+#include "RecommendedLayer.hpp"
+#include "../Utils.hpp"
+#include "../RecommendedUtils.hpp"
+
+//geode namespace
+using namespace geode::prelude;
+
+void RecommendedLayer::keyBackClicked() {
+	CCDirector::sharedDirector()->popSceneWithTransition(0.5f, PopTransition::kPopTransitionFade);
+};
+
+void RecommendedLayer::backButton(CCObject* sender) {
+	keyBackClicked();
+};
+
+bool RecommendedLayer::init() {
+	if (!CCLayer::init()) return false;
+
+	auto data = Mod::get()->getSavedValue<matjson::Value>("cached-data");
+
+	auto director = CCDirector::sharedDirector();
+	auto size = director->getWinSize();
+
+	auto bg = createLayerBG();
+	if (!Mod::get()->getSettingValue<bool>("restore-bg-color")) {
+		bg->setColor({ 18, 18, 86 });
+	}
+	bg->setZOrder(-10);
+	bg->setID("bg");
+	this->addChild(bg);
+
+	auto lCornerSprite = CCSprite::createWithSpriteFrameName("GJ_sideArt_001.png");
+	lCornerSprite->setAnchorPoint({ 0, 0 });
+	lCornerSprite->setID("left-corner-sprite");
+	this->addChild(lCornerSprite);
+
+	auto rCornerSprite = CCSprite::createWithSpriteFrameName("GJ_sideArt_001.png");
+	rCornerSprite->setAnchorPoint({ 1, 0 });
+	rCornerSprite->setPosition({ size.width, 0 });
+	rCornerSprite->setFlipX(true);
+	rCornerSprite->setID("right-corner-sprite");
+	this->addChild(rCornerSprite);
+
+	//error text
+	m_errorText = CCLabelBMFont::create("Something went wrong...", "bigFont.fnt");
+	m_errorText->setAlignment(CCTextAlignment::kCCTextAlignmentCenter);
+	m_errorText->setPosition({ size.width / 2, size.height / 2 });
+	m_errorText->setScale(0.6f);
+	m_errorText->setZOrder(100);
+	m_errorText->setVisible(false);
+	m_errorText->setID("error-text");
+	this->addChild(m_errorText);
+
+	//back button
+	auto backSprite = CCSprite::createWithSpriteFrameName("GJ_arrow_01_001.png");
+	auto backButton = CCMenuItemSpriteExtra::create(backSprite, this, menu_selector(RecommendedLayer::backButton));
+	auto backMenu = CCMenu::create();
+	backMenu->addChild(backButton);
+	backMenu->setPosition({ 25, size.height - 25 });
+	backMenu->setZOrder(2);
+	backMenu->setID("back-menu");
+	this->addChild(backMenu);
+
+	//info button
+	auto infoMenu = CCMenu::create();
+	auto infoButton = InfoAlertButton::create("Recommendations", "Levels that have been selected through an algorithm to give you the best options to improve your skills.", 1.0f);
+	infoMenu->setPosition({ 25, 25 });
+	infoMenu->setZOrder(2);
+	infoMenu->addChild(infoButton);
+	infoMenu->setID("info-menu");
+	this->addChild(infoMenu);
+
+	//reload menu
+	auto reloadMenu = CCMenu::create();
+	reloadMenu->setPosition({ size.width - 30, size.height - 30 });
+	auto reloadBtnSprite = CCSprite::createWithSpriteFrameName("GJ_updateBtn_001.png");
+	auto reloadBtn = CCMenuItemSpriteExtra::create(reloadBtnSprite, this, menu_selector(RecommendedLayer::reloadLevels));
+	reloadBtn->setPosition({ 0, 0 });
+	reloadMenu->addChild(reloadBtn);
+	reloadMenu->setID("reload-menu");
+	this->addChild(reloadMenu);
+
+	//pages menu
+	m_pagesMenu = CCMenu::create();
+	m_pagesMenu->setPosition({ 0, 0 });
+	m_pagesMenu->setID("pages-menu");
+
+	m_left = CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("GJ_arrow_03_001.png"), this, menu_selector(RecommendedLayer::pageLeft));
+	m_left->setPosition(24.f, size.height / 2);
+	m_left->setVisible(false);
+	m_pagesMenu->addChild(m_left);
+
+	auto rightBtnSpr = CCSprite::createWithSpriteFrameName("GJ_arrow_03_001.png");
+	rightBtnSpr->setFlipX(true);
+	m_right = CCMenuItemSpriteExtra::create(rightBtnSpr, this, menu_selector(RecommendedLayer::pageRight));
+	m_right->setPosition(size.width - 24.0f, size.height / 2);
+	m_right->setVisible(false);
+	m_pagesMenu->addChild(m_right);
+
+	this->addChild(m_pagesMenu);
+
+	m_list = GJListLayer::create(CustomListView::create(CCArray::create(), BoomListType::Level, 220.0f, 358.0f), "", { 194, 114, 62, 255 }, 358.0f, 220.0f, 0);
+	m_list->setZOrder(2);
+	m_list->setPosition(size / 2 - m_list->getContentSize() / 2);
+	this->addChild(m_list);
+
+	//check for errors
+	auto jsonCheck = JsonChecker(data);
+
+	if (jsonCheck.isError()) {
+		auto alert = FLAlertLayer::create("ERROR", fmt::format("Something went wrong validating the list data. ({})", jsonCheck.getError()), "OK");
+		alert->setParent(this);
+		alert->show();
+
+		return true;
+	}
+
+	loadLevels(0);
+
+	this->setKeyboardEnabled(true);
+	this->setKeypadEnabled(true);
+
+	return true;
+}
+
+void RecommendedLayer::reloadLevels(CCObject* sender) {
+	m_errorText->setVisible(false);
+
+	if (m_levelsLoaded) {
+		loadLevels(m_page);
+	}
+
+	return;
+}
+
+void RecommendedLayer::pageLeft(CCObject* sender) {
+	m_page -= 1;
+
+	loadLevels(m_page);
+
+	return;
+}
+
+void RecommendedLayer::pageRight(CCObject* sender) {
+	m_page += 1;
+
+	loadLevels(m_page);
+
+	return;
+}
+
+void RecommendedLayer::loadLevels(int page) {
+
+	m_pagesMenu->setVisible(false);
+
+	m_levelsLoaded = false;
+
+	m_loadCircle = LoadingCircle::create();
+	m_loadCircle->m_parentLayer = this;
+	m_loadCircle->show();
+
+	m_list->m_listView->setVisible(false);
+
+	matjson::Array levelIDs = Mod::get()->getSavedValue<matjson::Array>("recommended-levels");
+		
+	m_IDs.clear();
+
+	for (auto const& level : levelIDs) {
+		m_IDs.push_back(std::to_string(level.as_int()));
+	}
+
+	log::info("{}", m_IDs);
+
+	m_right->setVisible(m_IDs.size() > 10);
+
+	//borrowed some stuff from integrated demon list
+	auto glm = GameLevelManager::sharedState();
+	glm->m_levelManagerDelegate = this;
+	auto results = std::vector<std::string>(m_IDs.begin() + m_page * 10,
+		m_IDs.begin() + std::min(static_cast<int>(m_IDs.size()), (m_page + 1) * 10));
+	auto searchObject = GJSearchObject::create(SearchType::Type19, string::join(results, ","));
+	auto storedLevels = glm->getStoredOnlineLevels(searchObject->getKey());
+
+	//log::info("{}", searchObject);
+	//log::info("{}", searchObject->getKey());
+	//log::info("{}", storedLevels);
+
+	if (storedLevels) {
+		loadLevelsFinished(storedLevels, "");
+	}
+	else
+	{
+		glm->getOnlineLevels(searchObject);
+	}
+
+	return;
+}
+
+void RecommendedLayer::loadLevelsFinished(CCArray* levels, const char*) {
+
+	auto listSize = m_IDs.size();
+	auto maxPage = (listSize % 10 == 0 ? listSize : listSize + (10 - (listSize % 10))) / 10 - 1;
+	m_left->setVisible(m_page > 0);
+	m_right->setVisible(m_page < maxPage);
+	m_pagesMenu->setVisible(true);
+
+	m_levelsLoaded = true;
+	m_list->m_listView->setVisible(true);
+
+	m_loadCircle->fadeAndRemove();
+
+	auto director = CCDirector::sharedDirector();
+	auto size = director->getWinSize();
+
+	if (m_list->getParent() == this) { this->removeChild(m_list); }
+
+	m_list = GJListLayer::create(CustomListView::create(levels, BoomListType::Level, 220.0f, 358.0f), "Recommendations", {194, 114, 62, 255}, 358.0f, 220.0f, 0);
+	m_list->setZOrder(2);
+	m_list->setPosition(size / 2 - m_list->getContentSize() / 2);
+	this->addChild(m_list);
+
+	return;
+}
+
+void RecommendedLayer::loadLevelsFailed(const char*) {
+	m_levelsLoaded = true;
+
+	m_loadCircle->fadeAndRemove();
+
+	/*auto alert = FLAlertLayer::create("ERROR", "Failed to load levels. Please try again later.", "OK");
+	alert->setParent(this);
+	alert->show();*/
+
+	m_errorText->setVisible(true);
+
+	return;
+}
+
+RecommendedLayer* RecommendedLayer::create() {
+	auto pRet = new RecommendedLayer();
+	if (pRet && pRet->init()) {
+		pRet->autorelease();
+		return pRet;
+	}
+	CC_SAFE_DELETE(pRet); //don't crash if it fails
+	return nullptr;
+}
+
+RecommendedLayer::~RecommendedLayer() {
+	this->removeAllChildrenWithCleanup(true);
+}